% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediapackage_operations.R
\name{delete_channel}
\alias{delete_channel}
\title{Deletes an existing Channel}
\usage{
delete_channel(Id)
}
\arguments{
\item{Id}{The ID of the Channel to delete.}
}
\description{
Deletes an existing Channel.
}
\examples{

}
