% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediapackage_operations.R
\name{delete_origin_endpoint}
\alias{delete_origin_endpoint}
\title{Deletes an existing OriginEndpoint}
\usage{
delete_origin_endpoint(Id)
}
\arguments{
\item{Id}{The ID of the OriginEndpoint to delete.}
}
\description{
Deletes an existing OriginEndpoint.
}
\examples{

}
