% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediapackage_operations.R
\name{describe_channel}
\alias{describe_channel}
\title{Gets details about a Channel}
\usage{
describe_channel(
  Id = "string"
)
}
\arguments{
\item{Id}{[required] The ID of a Channel.}
}
\description{
Gets details about a Channel.
}
\examples{

}
