% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediapackage_operations.R
\name{describe_origin_endpoint}
\alias{describe_origin_endpoint}
\title{Gets details about an existing OriginEndpoint}
\usage{
describe_origin_endpoint(Id)
}
\arguments{
\item{Id}{[required] The ID of the OriginEndpoint.}
}
\description{
Gets details about an existing OriginEndpoint.
}
\section{Accepted Parameters}{
\preformatted{describe_origin_endpoint(
  Id = "string"
)
}
}

