% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediapackage_operations.R
\name{list_channels}
\alias{list_channels}
\title{Returns a collection of Channels}
\usage{
list_channels(
  MaxResults = 123,
  NextToken = "string"
)
}
\arguments{
\item{MaxResults}{Upper bound on number of records to return.}

\item{NextToken}{A token used to resume pagination from the end of a previous request.}
}
\description{
Returns a collection of Channels.
}
\examples{

}
