% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediapackage_operations.R
\name{list_origin_endpoints}
\alias{list_origin_endpoints}
\title{Returns a collection of OriginEndpoint records}
\usage{
list_origin_endpoints(ChannelId = NULL, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{ChannelId}{When specified, the request will return only OriginEndpoints associated with the given Channel ID.}

\item{MaxResults}{The upper bound on the number of records to return.}

\item{NextToken}{A token used to resume pagination from the end of a previous request.}
}
\description{
Returns a collection of OriginEndpoint records.
}
\examples{

}
