% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediapackage_package.R
\docType{package}
\name{paws.mediapackage-package}
\alias{paws.mediapackage}
\alias{paws.mediapackage-package}
\title{paws.mediapackage: AWS Elemental MediaPackage}
\description{
AWS Elemental MediaPackage
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
