% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediapackage_operations.R
\name{rotate_channel_credentials}
\alias{rotate_channel_credentials}
\title{Changes the Channel's first IngestEndpoint's username and password}
\usage{
rotate_channel_credentials(Id)
}
\arguments{
\item{Id}{The ID of the channel to update.}
}
\description{
Changes the Channel's first IngestEndpoint's username and password. WARNING - This API is deprecated. Please use RotateIngestEndpointCredentials instead
}
\examples{

}
