% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediapackage_operations.R
\name{rotate_ingest_endpoint_credentials}
\alias{rotate_ingest_endpoint_credentials}
\title{Rotate the IngestEndpoint's username and password, as specified by the IngestEndpoint's id}
\usage{
rotate_ingest_endpoint_credentials(IngestEndpointId, Id)
}
\arguments{
\item{IngestEndpointId}{The id of the IngestEndpoint whose credentials should be rotated}

\item{Id}{The ID of the channel the IngestEndpoint is on.}
}
\description{
Rotate the IngestEndpoint's username and password, as specified by the IngestEndpoint's id.
}
\examples{

}
