% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediapackage_operations.R
\name{update_channel}
\alias{update_channel}
\title{Updates an existing Channel}
\usage{
update_channel(Id, Description = NULL)
}
\arguments{
\item{Id}{The ID of the Channel to update.}

\item{Description}{A short text description of the Channel.}
}
\description{
Updates an existing Channel.
}
\examples{

}
