% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediapackage_operations.R
\name{create_channel}
\alias{create_channel}
\title{Creates a new Channel}
\usage{
create_channel(Id, Description = NULL)
}
\arguments{
\item{Id}{The ID of the Channel. The ID must be unique within the region and it
cannot be changed after a Channel is created.}

\item{Description}{A short text description of the Channel.}
}
\description{
Creates a new Channel.
}
\examples{

}
