% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediapackage_operations.R
\name{update_channel}
\alias{update_channel}
\title{Updates an existing Channel}
\usage{
update_channel(Description = NULL, Id)
}
\arguments{
\item{Description}{A short text description of the Channel.}

\item{Id}{[required] The ID of the Channel to update.}
}
\description{
Updates an existing Channel.
}
\section{Accepted Parameters}{
\preformatted{update_channel(
  Description = "string",
  Id = "string"
)
}
}

